Updated to use the master branch
Well, Elm has been upgraded to version 0.19 (this is quite a while ago).
However, This project uses dependencies that has not been updated (the
elm-mdl library). As a consequence I was not able to easily rebuild the
project. There, I simple changed all occurences of gh-pages to master
without rebuilding. The structure of the project is suboptimal. This
is something I do not intend to make better.